+++ flowman-core/src/main/scala/com/dimajix/flowman/jdbc/SqlExpressions.scala
@@ -25,4 +25,6 @@ abstract class SqlExpressions {
     def eq(column: String, value: Any): String
 
     def partition(partition: PartitionSpec): String
+
+    def collate(charset:Option[String], collation:Option[String]) : String
 }